refactor: convert NavigationSync component to useNavigationSync hook
Replace null-rendering component with custom hook for better semantics and performance. The hook provides the same navigation synchronization behavior while following React best practices.
